requirements.txt
streamlit
google-generativeai
python-dotenv
os
pillow